AdD utilities for building graph structures from WDL (#111)
Add utilities for building graph structures from WDL

+++ src/main/scala/wdlTools/syntax/v2/ConcreteSyntax.scala
@@ -35,7 +35,7 @@ object ConcreteSyntax {
   case class ExprNone(loc: SourceLocation) extends Expr
   case class ExprString(value: String, loc: SourceLocation) extends Expr
   case class ExprBoolean(value: Boolean, loc: SourceLocation) extends Expr
-  case class ExprInt(value: Int, loc: SourceLocation) extends Expr
+  case class ExprInt(value: Long, loc: SourceLocation) extends Expr
   case class ExprFloat(value: Double, loc: SourceLocation) extends Expr
 
   // represents strings with interpolation.
@@ -101,7 +101,7 @@ object ConcreteSyntax {
   sealed trait MetaValue extends Element
   case class MetaValueNull(loc: SourceLocation) extends MetaValue
   case class MetaValueBoolean(value: Boolean, loc: SourceLocation) extends MetaValue
-  case class MetaValueInt(value: Int, loc: SourceLocation) extends MetaValue
+  case class MetaValueInt(value: Long, loc: SourceLocation) extends MetaValue
   case class MetaValueFloat(value: Double, loc: SourceLocation) extends MetaValue
   case class MetaValueString(value: String, loc: SourceLocation) extends MetaValue
   case class MetaValueObject(value: Vector[MetaKV], loc: SourceLocation) extends MetaValue
